refactor to ExecutorServiceProvider (factory) (instead having the responsibility to create a new one on its own); refactor try catch with autoclosable try()

+++ libs/tacviewfx/src/main/java/com/recom/tacview/engine/TacViewer.java
@@ -93,7 +93,7 @@ public class TacViewer extends Canvas {
 
         // HANDLE TIME
         final long currentNanoTime = System.nanoTime();
-        final double elapsedNanoTime = (currentNanoTime - profiler.previousTickNanoTime);
+        final long elapsedNanoTime = (currentNanoTime - profiler.previousTickNanoTime);
         final long deltaTickNanoTime = currentNanoTime - profiler.previousTickNanoTime;
         final long deltaFrameNanoTime = currentNanoTime - profiler.previousFrameNanoTime;
