+++ libs/tacviewfx/src/main/java/com/recom/tacview/engine/module/DefaultEngineModule.java
@@ -52,7 +52,7 @@ public class DefaultEngineModule extends EngineModuleTemplate {
     }
 
     @Override
-    public void update(final double elapsedNanoTime) {
+    public void update(final long elapsedNanoTime) {
         super.update(elapsedNanoTime);
     }
